MedicalRecord entity, controller and service tested and refactored. tests improved

+++ src/main/java/com/safetynet/safetynetalerts/service/MedicalRecordServiceImpl.java
@@ -1,9 +1,13 @@
 package com.safetynet.safetynetalerts.service;
 
+import com.safetynet.safetynetalerts.DTOs.MedicalRecordDTO;
+import com.safetynet.safetynetalerts.convertor.MedicalRecordConverter;
 import com.safetynet.safetynetalerts.domain.MedicalRecord;
 import com.safetynet.safetynetalerts.repository.MedicalRecordRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
 import java.util.List;
@@ -14,15 +18,18 @@ public class MedicalRecordServiceImpl implements MedicalRecordService {
 
     //inject MedicalRecordRepository
     private MedicalRecordRepository medicalRecordRepository;
+    private MedicalRecordConverter medicalRecordConverter;
 
     @Autowired
-    public MedicalRecordServiceImpl(MedicalRecordRepository medicalRecordRepository) {
+    public MedicalRecordServiceImpl(MedicalRecordRepository medicalRecordRepository, MedicalRecordConverter medicalRecordConverter) {
         this.medicalRecordRepository = medicalRecordRepository;
+        this.medicalRecordConverter = medicalRecordConverter;
     }
 
     @Override
-    public List<MedicalRecord> findAll() {
-        return medicalRecordRepository.findAll();
+    public List<MedicalRecordDTO> findAll() {
+        List<MedicalRecord> medicalRecords = medicalRecordRepository.findAll();
+        return medicalRecordConverter.medicalRecordToDAOsConverter(medicalRecords);
     }
 
     @Override
@@ -32,7 +39,14 @@ public class MedicalRecordServiceImpl implements MedicalRecordService {
 
     @Override
     public void deleteByFirstNameAndLastName(String firstName, String lastName) {
-        medicalRecordRepository.deleteByFirstNameAndLastName(firstName, lastName);
+        MedicalRecord medicalRecord = medicalRecordRepository.findByFirstNameAndLastName(firstName, lastName);
+
+        if (medicalRecord == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No medical record that belongs to '" +
+                    firstName + " " + lastName + "' was found");
+        } else {
+            medicalRecordRepository.deleteByFirstNameAndLastName(firstName, lastName);
+        }
     }
 
     @Override
